Add support for klipper flavored gcode
When testing p2pp with Superslicer set to klipper flavored gcode, p2pp
fails to recognize the gcode as a multi-color print, because the gcode
for switching extruders is different than normal marlin/etc.  This adds
support for the klipperr `ACTIVATE_EXTRUDER` command that is the
equivalent of the T0/Tx commands.